disable libgeocoding
The feature is not needed at the moment and fails to build on armv7l
with random errors. For example:

[  217s] /home/abuild/rpmbuild/BUILD/libphonenumber/cpp/src/phonenumbers/geocoding/geocoding_data.cc:29686:25: error: expected primary-expression at end of input
[  217s]    "St. Georgen am Walde",
[  217s]                          ^

The randomness may come from the code generator which produces the
geocoding_data.cc file. However, the file itself is okay and compiles
fine with a gcc from a different architecture. Looks like a problem
with the arm toolchain.

Disabling the compilation of libgeocoding is the fastest way of
solving this, which is acceptable because it is not needed at the
moment. Should it ever be needed, perhaps a different version of the
toolchain at that time will be able to handle it.

Change-Id: Ibd3c1414ad2f7842dfa5af474ef6d4f9f4adbb57
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>